Change some Jitsi rooms and external links
Remove links to GitHub and streaming.media.ccc.de (GitHub is a non-EU
service provider and thus not acceptable and the world is not really
intended to watch the streams). Also use only two Jitsi rooms, one of
which 'magically' connects several areas (by using the same room
identifier). In addition to this, also add a few more links to tag
descriptions in the OSM wiki to have some more interactivity.